Addressed miscallaneous feedback from @albrow
