+++ orderfilter/validate_js.go
@@ -3,16 +3,11 @@
 package orderfilter
 
 import (
-	"context"
 	"errors"
-	"fmt"
 	"syscall/js"
 
 	"github.com/0xProject/0x-mesh/packages/browser/go/jsutil"
 	"github.com/0xProject/0x-mesh/zeroex"
-	peer "github.com/libp2p/go-libp2p-core/peer"
-	pubsub "github.com/libp2p/go-libp2p-pubsub"
-	log "github.com/sirupsen/logrus"
 )
 
 type SchemaValidationError struct {
@@ -54,7 +49,7 @@ func (f *Filter) ValidateOrderJSON(orderJSON []byte) (*SchemaValidationResult, e
 	if !f.checkForValidator() {
 		return nil, errors.New(`"schemaValidator" has not been set on the "window" object`)
 	}
-	jsResult := js.Global().Get("schemaValidator").Call("orderValidator", js.ValueOf(string(orderJSON)))
+	jsResult := js.Global().Get("schemaValidator").Call("orderValidator", string(orderJSON))
 	fatal := jsResult.Get("fatal")
 	if !jsutil.IsNullOrUndefined(fatal) {
 		return nil, errors.New(fatal.String())
@@ -63,20 +58,9 @@ func (f *Filter) ValidateOrderJSON(orderJSON []byte) (*SchemaValidationResult, e
 	jsErrors := jsResult.Get("errors")
 	var convertedErrors []*SchemaValidationError
 	for i := 0; i < jsErrors.Length(); i++ {
-		convertedErrors = append(convertedErrors, &SchemaValidationError{errors.New(jsErrors.Get(fmt.Sprintf("%d", i)).String())})
+		convertedErrors = append(convertedErrors, &SchemaValidationError{errors.New(jsErrors.Index(i).String())})
 	}
 	return &SchemaValidationResult{valid: valid, errors: convertedErrors}, nil
-}
-
-// MatchOrder returns true if the order passes the filter. It only returns an
-// error if there was a problem with validation. For details about
-// orders that do not pass the filter, use ValidateOrder.
-func (f *Filter) MatchOrder(order *zeroex.SignedOrder) (bool, error) {
-	result, err := f.ValidateOrder(order)
-	if err != nil {
-		return false, err
-	}
-	return result.Valid(), nil
 }
 
 func (f *Filter) MatchOrderMessageJSON(messageJSON []byte) (bool, error) {
@@ -97,19 +81,4 @@ func (f *Filter) ValidateOrder(order *zeroex.SignedOrder) (*SchemaValidationResu
 		return nil, err
 	}
 	return f.ValidateOrderJSON(orderJSON)
-}
-
-// Dummy declaration to ensure that ValidatePubSubMessage matches the expected
-// signature for pubsub.Validator.
-var _ pubsub.Validator = (&Filter{}).ValidatePubSubMessage
-
-// ValidatePubSubMessage is an implementation of pubsub.Validator and will
-// return true if the contents of the message pass the message JSON Schema.
-func (f *Filter) ValidatePubSubMessage(ctx context.Context, sender peer.ID, msg *pubsub.Message) bool {
-	isValid, err := f.MatchOrderMessageJSON(msg.Data)
-	if err != nil {
-		log.WithError(err).Error("MatchOrderMessageJSON returned an error")
-		return false
-	}
-	return isValid
 }
